validations added

+++ src/main/java/com/rest/user/resource/UsersResource.java
@@ -26,14 +26,18 @@ public class UsersResource {
     }
 
     @PutMapping(value = "/user")            // put mapping
-    public List<Users> update(@RequestBody final Users users) {
-        usersRepository.save(users);
-        return usersRepository.findAll();
+    public void update(@RequestBody final Users users ,@RequestParam(value = "userId", defaultValue = "") String id) {
+        if(usersRepository.findById(id).get() != null) {
+            usersRepository.save(users);
+            //return usersRepository.findAll();
+        }
     }
 
     @DeleteMapping(value = "/user")         // delete mapping
     public void deleteUser(@RequestParam(value = "userId", defaultValue = "") String id) {
-        usersRepository.deleteById(id);
+        if(usersRepository.findById(id).get() != null) {
+            usersRepository.deleteById(id);
+        }
     }
 
 }
